requirements.txt
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
PyYAML>=6.0.1
ccxt>=4.2.0
streamlit>=1.30.0
plotly>=5.18.0
pytest>=7.4.0
